+++ src/main/java/com/banquito/core/banking/seguridadbanco/dao/RolRepository.java
@@ -1,5 +1,6 @@
 package com.banquito.core.banking.seguridadbanco.dao;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
@@ -8,8 +9,8 @@ import org.springframework.stereotype.Repository;
 import com.banquito.core.banking.seguridadbanco.domain.Rol;
 
 @Repository
-public interface RolRepository extends CrudRepository<Rol, Integer> {
-    public List<Rol> findByCodRolOrderByCodRol(Integer codRol);
+public interface RolRepository extends CrudRepository<Rol, BigDecimal> {
+    public List<Rol> findByCodRolOrderByCodRol(BigDecimal codRol);
     public List<Rol> findByNombreRolAndResponsableOrderByFechaCreacion(String nombreRol, String responsable);
-    public List<Rol> findByCodRolAndResponsableOrderByNombreRol(Integer codRol, String responsable);
+    public List<Rol> findByCodRolAndResponsableOrderByNombreRol(BigDecimal codRol, String responsable);
 }
